Updates `trainers/*`
